requirements.txt
Flask==3.1.1
moviepy==2.2.1
python-dotenv==1.1.1
imageio==2.37.0
imageio-ffmpeg==0.6.0
numpy==2.3.1
pillow==11.3.0
proglog==0.1.12
tqdm==4.67.1
requests
pydub
